+++ build.gradle
@@ -33,6 +33,10 @@ dependencies {
     implementation 'com.h2database:h2:2.3.232'
 //    implementation 'io.awspring.cloud:spring-cloud-aws-starter-parameter-store:3.3.0'
     implementation 'org.modelmapper:modelmapper-module-record:1.0.0'
+    implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
+    implementation 'org.thymeleaf.extras:thymeleaf-extras-springsecurity6'
+    implementation 'nz.net.ultraq.thymeleaf:thymeleaf-layout-dialect' //:3.4.0'
+    implementation 'org.webjars:bootstrap:5.3.3'
 
     compileOnly 'org.projectlombok:lombok'
