+++ src/main/java/com/ageinghippy/controller/advice/GutControllerAdvice.java
@@ -6,10 +6,11 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 import java.util.NoSuchElementException;
 
-@ControllerAdvice
+@RestControllerAdvice(basePackages = {"com.ageinghippy.controller.api"})
 @Slf4j
 public class GutControllerAdvice {
