+++ src/main/resources/templates/fragments/layout.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+      xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="utf-8"/>
+    <meta http-equiv="X-UA-Compatible" content="IE=edge"/>
+    <meta name="viewport" content="width=device-width, initial-scale=1"/>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css">
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
+
+    <title layout:title-pattern="$LAYOUT_TITLE - $CONTENT_TITLE">gutHealth</title>
+</head>
+<body>
+<th:block th:replace="~{fragments/header :: header}"/>
+<div class="container">
+    <th:block layout:fragment="content"/>
+</div>
+<th:block th:replace="~{fragments/footer :: footer}"/>
+</body>
+</html>
