+++ src/main/resources/templates/index.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+      xmlns:th="http://www.thymeleaf.org"
+      layout:decorate="~{fragments/layout}">
+<head>
+    <meta charset="UTF-8">
+    <title>Index</title>
+</head>
+<body>
+
+<div layout:fragment="content" th:remove="tag">
+
+    <div align="center">
+        <h1>gutHealth Index</h1>
+        <br/><br/>
+        <p>Welcome Dude or Dudette</p>
+        <br>
+    </div>
+
+</div>
+
+</body>
+</html>
