+++ src/main/resources/templates/preparation-technique-edit.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+      xmlns:th="http://www.thymeleaf.org"
+      layout:decorate="~{fragments/layout}">
+<head>
+    <meta charset="UTF-8">
+    <title>Edit Preparation Technique</title>
+</head>
+<body>
+
+<div layout:fragment="content" th:remove="tag">
+
+    <div>
+        <form action="#" th:action="@{'/preparation-technique/update/' + ${preparationTechnique.code}}"
+              th:object="${preparationTechnique}" method="post">
+            <div class="mb-3">
+                <label for="code" class="col-form-label">Code:</label>
+                <input type="text" class="form-control" id="code" th:field="*{code}" readonly>
+            </div>
+            <div class="mb-3">
+                <label for="description" class="col-form-label">Description:</label>
+                <input type="text" class="form-control" id="description" th:field="*{description}">
+            </div>
+            <button type="submit">Save</button>
+        </form>
+    </div>
+
+
+</div>
+
+</body>
+</html>
